requirements.txt
absl-py==0.7.1
appdirs==1.4.3
astor==0.8.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
comet-git-pure==0.19.11
comet-ml==2.0.5
configobj==5.0.6
coverage==4.5.3
cycler==0.10.0
dill==0.2.9
docopt==0.6.2
docopt-subcommands==3.0.0
everett==1.0.2
gast==0.2.2
giganticode-dataprep==1.0.0a6
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
idna==2.8
jsons==0.8.3
jsonschema==3.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.0.3
mock==3.0.5
netifaces==0.10.9
nltk==3.4.4
numpy==1.16.4
nvidia-ml-py3==7.352.0
protobuf==3.9.0
Pygments==2.3.1
pyparsing==2.4.1
pyrsistent==0.15.3
python-dateutil==2.8.0
PyYAML==5.1
regex==2019.3.12
requests==2.22.0
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tqdm==4.31.1
urllib3==1.25.3
websocket-client==0.56.0
Werkzeug==0.15.5
wrapt==1.11.2
wurlitzer==1.0.3
